// apps/api/src/modules/auth/interfaces/create-user.port.ts
export interface RegisterUserInput {
  email: string;
  anonName: string;
  passwordHash: string;
  emailVerificationTokenHash: string;
  emailVerificationTokenExpiresAt: Date;
}

export interface RegisterUserResult {
  id: string;
  anonName: string;
}

export interface RegisterUserPort {
  execute(dto: RegisterUserInput): Promise<RegisterUserResult>;
}

export const REGISTER_USER_PORT = Symbol('REGISTER_USER_PORT');